tests/claude/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-mock==3.12.0
respx==0.20.2
Pillow==10.1.0